backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-jwt-extended==4.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.26.0
apscheduler==3.8.1
gunicorn==20.1.0
